% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_function.R
\name{CostFun}
\alias{CostFun}
\title{Cost function for MDLM}
\usage{
CostFun(bs = b.start, ybar = "y", xbar, data, rmat = r)
}
\arguments{
\item{bs}{Intial beta}

\item{ybar}{Dependent variable.}

\item{xbar}{Independent variable.}

\item{data}{Data.}

\item{rmat}{Correlation matrix.}
}
\value{
Sum of squared error for MDLM
}
\description{
Least square estimation for MDLM
}
\examples{
# ADD_EXAMPLES_HERE
}
