% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlm.R
\docType{data}
\name{exMDLM}
\alias{exMDLM}
\title{This is data to be included in MDLM}
\format{An object of class \code{data.frame} with 1005 rows and 4 columns.}
\usage{
exMDLM
}
\description{
This is data to be included in MDLM
}
\references{
\url{https://github.com/jinseob2kim/MDLM}
}
\author{
Jinseob Kim \email{jinseob2kim@gmail.com}
}
\keyword{datasets}
