% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_function.R
\name{IniBetaMake}
\alias{IniBetaMake}
\title{Generate initial beta}
\usage{
IniBetaMake(ybar, xbar, data)
}
\arguments{
\item{ybar}{Dependent variable.}

\item{xbar}{Independent variable.}

\item{data}{Data.}
}
\value{
Initial beta
}
\description{
Get initial beta from conventional linear model
}
\examples{
# ADD_EXAMPLES_HERE
}
