% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_function.R
\name{LowerToMatrix}
\alias{LowerToMatrix}
\title{Convert lower triangle to Correlation matrix}
\usage{
LowerToMatrix(lower = c(4, 7, 8))
}
\arguments{
\item{lower}{Lower triangle. The length is must be x(x-1)/2 form}
}
\value{
Matrix
}
\description{
Convert lower triangle to Correlation matrix
}
\examples{
# ADD_EXAMPLES_HERE
}
