% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlm.R
\name{mdlm1}
\alias{mdlm1}
\title{MDLM of fixed vector space}
\usage{
mdlm1(ybar, xbar, data, cor.mat = NULL, ini.beta = NULL)
}
\arguments{
\item{ybar}{Dependent variable.}

\item{xbar}{Independent variable.}

\item{data}{Data.}

\item{cor.mat}{Dependency matrix of vector space}

\item{ini.beta}{Initial beta.}
}
\value{
beta, standard error, z value, p value, rmse
}
\description{
MDLM of fixed vector space
}
\examples{
mdlm1(ybar="y",xbar=c("x1","x2"),data=exMDLM, cor.mat=NULL, ini.beta=NULL)
}
