% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_function.R
\name{SqrtFun}
\alias{SqrtFun}
\title{Square sum for CostFun}
\usage{
SqrtFun(bs = b.start, ybar = "y", xbar, data, rmat = r)
}
\arguments{
\item{bs}{Intial beta}

\item{ybar}{Dependent variable.}

\item{xbar}{Independent variable.}

\item{data}{Data.}

\item{rmat}{Correlation matrix.}
}
\value{
Squared sum value
}
\description{
Square sum for CostFun
}
\examples{

}
