% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlm.R
\name{mdlm2}
\alias{mdlm2}
\title{MDLM of non-fixed vector space}
\usage{
mdlm2(ybar, xbar, data, ini.cor = 0, ini.beta = NULL)
}
\arguments{
\item{ybar}{Dependent variable.}

\item{xbar}{Independent variable.}

\item{data}{Data.}

\item{ini.cor}{Initial g_ij for all g_ij}

\item{ini.beta}{Initial beta}
}
\value{
beta, g_ij, standard error, z value, p value, rmse
}
\description{
Also estimate g_ij
}
\examples{
# ADD_EXAMPLES_HERE
}
